+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3misc (development version)
+# mlr3misc 0.23.0
 
 * fix: `hash_input()` for functions no longer loses the argument names of a body that is a single call.
 * feat: Added a `hash_input()` method for `list`s that applies `hash_input()` to all elements.